pyproject.toml
[tool.poetry]
name = "bss-web-file-server"
version = "0.1.0"
description = "BSS file server API"
authors = ["Bence Csik <bence@csik.dev>"]
readme = "README.md"
packages = [{include = "bss_web_file_server"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.87.0"
pillow = "9.3.0"
uvicorn = {extras = ["standard"], version = "0.19.0"}
python-multipart = "0.0.5"
pillow-avif-plugin = "1.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
pylint = "^2.15.5"
black = "^22.10.0"
isort = "^5.10.1"
bandit = "^1.7.4"

[tool.isort]
profile = "black"

[tool.bandit]
targets = ["app"]
exclude_dirs = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
